Update totp.py
Adds a date and time to the main screen of the TOTP authenticator. Comes at cost to some space for more TOTPs but I only have about 20 and this was dead space. May think of some other useful things to put on this data screen as I have this running basically all day, every day